+++ src/lib.rs
@@ -42,8 +42,8 @@
 
 pub mod client;
 pub mod error;
-mod to_socket_addr_with_hostname;
+mod to_socket_addrs_with_hostname;
 
 pub use client::{AuthMethod, Client, ServerCheckMethod};
 pub use error::Error;
-pub use to_socket_addr_with_hostname::ToSocketAddrsWithHostname;
+pub use to_socket_addrs_with_hostname::ToSocketAddrsWithHostname;
